# Conflicts:
#	src/Client/ViewModel/EditPersonalInfoViewModel.java
#	src/Client/Views/AddEmployeesView/AddEmployeesViewController.java
#	src/Client/Views/EditEmployeeInfoView/EditEmployeeInfoViewController.java
#	src/Client/Views/EditPersonalInfoView/EditPersonalInfoViewController.java

+++ src/Client/ViewModel/EditEmployeeInfoViewModel.java
@@ -18,7 +18,7 @@ public class EditEmployeeInfoViewModel
     this.userModel=userModel;
   }
   public void editEmployeeInfo(Employee employee,String cpr, String firstName, String lastName, GregorianCalendar dateOfBirth,String phoneNumber,
-      String email, int salary,String position)
+      String email, double salary,String position)
       throws RemoteException, SQLException
   {
     Employee newEmployee = new Employee(cpr,firstName,lastName,dateOfBirth,phoneNumber,email,salary,position);
